Mark as done

+++ src/components/TaskPopup.vue
@@ -1,8 +1,8 @@
 <template>
 	<div v-show="visible" class="dtask">
 		<div class="dtask__cont">
-			<label class="dtask__label">Enter the task id to be deleted: </label>
-			<input class="dtask__input" ref="delete" maxlength="2" v-model.trim="id">
+			<label class="dtask__label">{{label}}</label>
+			<input class="dtask__input" ref="input" maxlength="2" v-model.trim="id">
 		</div>
 		<CommandTips/>
 	</div>
@@ -13,7 +13,7 @@ import CommandTips from '@/components/CommandTips'
 import {EventBus} from '@/eventBus'
 import {Storage} from '@/storage'
 export default {
-	props: ['visible'],
+	props: ['visible', 'label'],
 	data() {
 		return {
 			id: ''
@@ -21,15 +21,36 @@ export default {
 	},
 	components: {CommandTips},
 	methods: {
-		deleteTask() {
+		updateTask(params) {
 			if (!this.id) return
-			Storage.deleteTask(this.id)
-			this.id = ''
-			this.$emit('update')
+
+			if (params.action == 'DELETE') {
+				Storage.deleteTask(this.id)
+				this.id = ''
+				this.$emit('update')
+				return
+			}
+
+			if (params.action == 'DONE') {
+				let task = Storage.getTask(this.id)
+				if (!task) {
+					this.id = ''
+					this.$emit('update')
+					return
+				}
+
+				let status = task.status === 'CREATED' ? 'DONE' : 'CREATED'
+				Storage.updateTask(this.id, status)
+				this.id = ''
+				this.$emit('update')
+				return
+			}
+
+			console.error(`Action ${params.action} not set`)
 		}
 	},
 	mounted() {
-		EventBus.$on('delete-task', () => this.deleteTask());
+		EventBus.$on('update-task', (params) => this.updateTask(params));
 	}
 }
 </script>
